+++ src/main/java/roomescape/auth/AuthController.java
@@ -1,0 +1,56 @@
+package roomescape.auth;
+
+import io.jsonwebtoken.Jwts;
+import io.jsonwebtoken.security.Keys;
+import jakarta.servlet.http.Cookie;
+import jakarta.servlet.http.HttpServletResponse;
+import org.springframework.http.ResponseEntity;
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import roomescape.member.Member;
+import roomescape.member.MemberDao;
+
+@Controller
+public class AuthController {
+
+    private final MemberDao memberDao;
+
+    public AuthController(MemberDao memberDao) {
+        this.memberDao = memberDao;
+    }
+
+    @PostMapping("/login")
+    public ResponseEntity login(@RequestBody LoginRequest loginRequest, HttpServletResponse response) {
+        String email = loginRequest.email();
+        String password = loginRequest.password();
+        Member findMember = memberDao.findByEmailAndPassword(email, password);
+        if (findMember == null) {
+            throw new RuntimeException("로그인 실패입니다.");
+        }
+
+        String secretKey = "Yn2kjibddFAWtnPJ2AFlL8WXmohJMCvigQggaEypa5E=";
+        String accessToken = Jwts.builder()
+                .setSubject(findMember.getId().toString())
+                .claim("name", findMember.getName())
+                .claim("role", findMember.getRole())
+                .signWith(Keys.hmacShaKeyFor(secretKey.getBytes()))
+                .compact();
+        Cookie cookie = new Cookie("token", accessToken);
+        cookie.setPath("/");
+        cookie.setHttpOnly(true);
+        response.addCookie(cookie);
+        return ResponseEntity.ok().build();
+    }
+
+    @PostMapping("/logout")
+    public ResponseEntity logout(HttpServletResponse response) {
+        Cookie cookie = new Cookie("token", "");
+        cookie.setHttpOnly(true);
+        cookie.setPath("/");
+        cookie.setMaxAge(0);
+        response.addCookie(cookie);
+        return ResponseEntity.ok().build();
+    }
+
+}
