+++ src/main/java/roomescape/auth/LoginRequest.java
@@ -1,0 +1,4 @@
+package roomescape.auth;
+
+public record LoginRequest(String email, String password) {
+}
